+++ src/Enclave/eth_transaction.h
@@ -34,11 +34,11 @@ extern "C" {
 #endif
 
 void rlp_item(const uint8_t*, const int, bytes&);
-int get_raw_signed_tx(int nonce, int nonce_len, 
+int get_raw_signed_tx(int nonce, int nonce_len,
                       uint64_t request_id, uint8_t request_type,
                       const uint8_t* req_data, int req_len,
                       bytes resp_data,
-                      uint8_t* serialized_tx, int* o_len);
+                      uint8_t *serialized_tx, size_t *o_len);
 
 #if defined(__cplusplus)
 }
